Try and provide a clear error message when pip installed directly, fix symlink farm issue, fix scripts issue, TODO: fix SPARK_HOME and find out why JARs aren't ending up in the install

+++ python/setup.py
@@ -8,20 +8,34 @@ TEMP_PATH = "deps"
 SPARK_HOME = os.path.abspath("../")
 JARS_PATH = "%s/assembly/target/scala-2.11/jars/" % SPARK_HOME
 SCRIPTS_PATH = "%s/bin" % SPARK_HOME
-SCRIPTS = "%s/bin" % TEMP_PATH
-JARS = "%s/jars" % TEMP_PATH
+SCRIPTS_TARGET = "%s/bin" % TEMP_PATH
+JARS_TARGET = "%s/jars" % TEMP_PATH
 
 
-# Construct links for setup
+# Check and see if we are under the spark path in which case we need to build the symlink farm.
+# The py4j src file is used to check this since for pip installed we use the py4j libraries rather
+# than the source zip.
+in_spark = os.path.isfile("lib/py4j-0.10.3-src.zip")
+if (in_spark):
+    # Construct links for setup
+    try:
+        os.mkdir(TEMP_PATH)
+    except:
+        print("Temp path for symlink to parent already exists %s" % TEMP_PATH, file=sys.stderr)
+        exit(-1)
+
 try:
-    os.mkdir(TEMP_PATH)
-except:
-    print("Temp path for symlink to parent already exists %s" % TEMP_PATH, file=sys.stderr)
-    exit(-1)
+    if (in_spark):
+        os.symlink(JARS_PATH, JARS_TARGET)
+        os.symlink(SCRIPTS_PATH, SCRIPTS_TARGET)
 
-try:
-    os.symlink(JARS_PATH, JARS)
-    os.symlink(SCRIPTS_PATH, SCRIPTS)
+    if not os.path.isdir(SCRIPTS_TARGET):
+        print("For packaging reasons you must first create a source dist and install that source dist.", file=sys.stderr)
+        exit(-1)
+
+    # Scripts directive requires a list of each script path and does not take wild cards.
+    script_names = os.listdir(SCRIPTS_TARGET)
+    scripts = map(lambda script: SCRIPTS_TARGET + "/" + script, script_names)
 
     setup(
         name='pyspark',
@@ -37,8 +51,8 @@ try:
                   'pyspark.streaming'],
         include_package_data=True,
         package_data={
-            'pyspark': [JARS]},
-        scripts=[SCRIPTS + "/*"],
+            'pyspark': [JARS_TARGET + "/*"]},
+        scripts=scripts,
         license='http://www.apache.org/licenses/LICENSE-2.0',
         install_requires=['py4j==0.10.3'],
         extras_require={
@@ -48,6 +62,9 @@ try:
         }
     )
 finally:
-    os.remove("%s/jars" % TEMP_PATH)
-    os.remove("%s/bin" % TEMP_PATH)
-    os.rmdir(TEMP_PATH)
+    # We only cleanup the symlink farm if we were in Spark, otherwise we are installing rather than
+    # packaging.
+    if (in_spark):
+        os.remove("%s/jars" % TEMP_PATH)
+        os.remove("%s/bin" % TEMP_PATH)
+        os.rmdir(TEMP_PATH)
